TNT tiles being added (still require a degree of integration with grid/player).
WARNING: I modified the hit function in unit.h by switching the placement of the 3rd and 4th parameters.  I noticed the new version of hit was (...damage, rin) whereas the old function was (...radius, damage).  I have made the function (damage, rin), assuming this is the correct way to make the function compatible with older hit calls.